Refactor "generate" matcher to "generateValues"
