+++ src/Extension.php
@@ -15,8 +15,8 @@ final class Extension implements ExtensionInterface
      */
     public function load(ServiceContainer $container)
     {
-        $container->set('matchers.generate', function (ServiceContainer $c) {
-            return new Matcher\GenerateMatcher($c->get('formatter.presenter'));
+        $container->set('matchers.generate_values', function (ServiceContainer $c) {
+            return new Matcher\GenerateValuesMatcher($c->get('formatter.presenter'));
         });
     }
 }
